Better error messages.
We were already checking for
(dFar <= dNear || dNear <= 0)
We would print "invalid" in the result if we saw this.
But now we also print an error message adjacent to the input.
Messages include "This value must be <u>greater than</u> 0."
and "The far value must be <u>greater than</u> the near value."